PYCOLL.ENH: refactor impedances module into a package.
